moves test js into compilation dir
