+++ test/deno/CompileTests.hs
@@ -4,5 +4,5 @@ import Compile (runCompilation)
 
 main :: IO ()
 main = do
-  runCompilation "test/deno/math.po" "test/deno/math.js"
-  runCompilation "test/deno/hof.po" "test/deno/hof.js"
+  runCompilation "test/deno/math.po" "test/deno/compiled/math.js"
+  runCompilation "test/deno/hof.po" "test/deno/compiled/hof.js"
