+++ test/deno/hof.test.js
@@ -1,6 +1,6 @@
 import _ from 'https://deno.land/x/lodash@4.17.15-es/lodash.js'
 import { assert, assertEquals } from "https://deno.land/std@0.119.0/testing/asserts.ts"
-import { compose, double, quadruple, id, id2, one1, one2 } from './hof.js'
+import { compose, double, quadruple, id, id2, one1, one2 } from './compiled/hof.js'
 
 const generateNumArray = () => Array(Math.floor(100 * Math.random()))
     .fill(0)
